requirements.txt
fastapi==0.115.12
ftfy==6.3.1
matplotlib==3.8.0
numpy
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==11.2.1
plotly==5.22.0
pydantic==2.11.5
pytorchcv==0.0.73
regex==2024.11.6
Requests==2.32.3
scikit_learn==1.4.1.post1
streamlit>=1.28.0
setuptools==68.2.2
torch==2.2.1
python-multipart
torchvision
tqdm==4.67.1
uvicorn==0.34.2

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2